+++ rx_main_ctrl/gui_svr.c
@@ -223,7 +223,7 @@ void gui_tick(void)
 	{
 		if (_CheckTime[i] || _CheckRecv[i] || _Sockets[i]!=INVALID_SOCKET) 
 		{
-			len += sprintf(&str[len], "gui[%d](sok=%d, r=%03d, s=%03d)", i, _Sockets[i], _CheckRecv[i], _CheckSend[i]);
+			len += sprintf(&str[len], "gui[%d](sok=%d, r=%03d, s=%03d)", i, (int)_Sockets[i], _CheckRecv[i], _CheckSend[i]);
 			if (_CheckTime[i] && time-_CheckTime[i]>TIMEOUT)
 			{
 				_TimeoutCnt++;
